+++ negotiation/src/test/java/it/eng/negotiation/policy/evaluator/AccessCountPolicyEvaluatorTest.java
@@ -70,7 +70,7 @@ class AccessCountPolicyEvaluatorTest {
 	                .description("Access count policy")
 	                .enabled(true)
 	                .validFrom(LocalDateTime.now().minusDays(1))
-	                .attribute(PolicyConstants.COUNT, 3) // Max count
+	                .attribute(PolicyConstants.COUNT, 4) // Max count
 	                .attribute(PolicyConstants.OPERATOR, Operator.LT) // Operator
 	                .build();
 		  
@@ -102,7 +102,7 @@ class AccessCountPolicyEvaluatorTest {
 	                .description("Access count policy")
 	                .enabled(true)
 	                .validFrom(LocalDateTime.now().minusDays(1))
-	                .attribute("count", 3) // Max count
+	                .attribute("count", 4) // Max count
 	                .attribute("operator", Operator.LT) // Operator
 	                .build();
 		  
@@ -175,7 +175,7 @@ class AccessCountPolicyEvaluatorTest {
 	                .agreementId("agreement-123")
 	                .resourceId("resource-123")
 	                .userId("user-123")
-	                .attribute(PolicyConstants.CURRENT_COUNT, 4) // Current count
+	                .attribute(PolicyConstants.CURRENT_COUNT, 3) // Current count
 	                .action(Action.ANONYMIZE)
 	                .build();